Moved everything to base. Split principia::std into a separate file. Disabled principia::std for non-macOS builds.
